requirements.txt
anyascii==0.3.2
asgiref==3.8.1
beautifulsoup4==4.12.3
certifi==2024.8.30
charset-normalizer==3.3.2
defusedxml==0.7.1
dj-database-url==2.2.0
Django==4.2.16
django-filter==24.3
django-modelcluster==6.3
django-permissionedforms==0.1
django-taggit==5.0.1
django-treebeard==4.7.1
django-widget-tweaks==1.5.0
djangorestframework==3.15.2
djmail==2.0.0
draftjs-exporter==5.0.0
et-xmlfile==1.1.0
filetype==1.2.0
idna==3.8
l18n==2021.3
laces==0.1.1
openpyxl==3.1.5
pillow==10.4.0
pillow_heif==0.18.0
psycopg[binary]==3.2.10
pytz==2024.1
requests==2.32.3
six==1.16.0
soupsieve==2.6
sqlparse==0.5.1
telepath==0.3.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
wagtail==6.2.1
wagtail-modeladmin==2.0.0
Willow==1.8.0
boto3==1.35.0
django-storages==1.14.2
python-dotenv==1.0.0
